Logging and header echoing
#What
Added logging for request handling to see incoming requests and set up the server to echo incoming request headers.

#Why
Stepping stone for echoing the body data of a request, as it was easier to just send the raw incoming data to test whether the header configuration worked properly for curl requests.

#How
Request handling now generates an HTTP header string and appends the incoming data as a body.

Change-Id: Ic8456c2e80c9ea218501885059e137e9b639d61f