Content Length Works
#What
Skeleton for body parsing, and content-length implementation

#Why
Curl requests would hang forever, as the HTTP header I had set up previously didn't have a content-length header. I've implemented one to make the request complete.

#How
The body is now assigned to its own string, which we then take the size of for the header.

Change-Id: Ib6e54651b2e861196f18f4f422105b5f63fe7066