return body
#What
Updated code to return the body of a request. Can test with:
`curl -X POST http://localhost:8080 -d "Hello World"`

#Why
We're supposed to be returning the body in the assignment guidelines.

#How
There's a data parsing function now that takes in the raw data and extracts the information.

#TODO
For some reason the response handler seems to be called twice, and errors out once each time - I do not know why as of now, but the code still works properly behavior-wise so I figured I should leave that to future commits.

Additionally, the data parsing function should eventually check for proper requests.



Change-Id: Ided3a36eb2247bf8f3351ea09688446037d2c7d8