nginx parsing
#NOTE:
I know this LOOKS like a miserably long commit - that's because this is mostly an import of my assignment 1, complete with the tests. The main thing to look at here is just in server_main.cc.

#What
Instead of a port, the server now takes in a path to a config file. Currently, the new launch command is by default:

`bin/server ../server_config`

as the config I created is in the main folder.

It uses port 8080 as that was the one we originally used in the assignment 2 instructions.

#Why
Assignment 2 requires servers to use config files.


#How
There is an admittedly janky parsing loop done in server_main.cc that extracts the config args to locate the port. The stuff I wrote is overboard for what we currently need, but is expandable compared to just directly hardcoding it in.

Change-Id: I948d41be6e21f3b62763e966f26e08a2eda718a2