Skeleton for request handlers and config parsing additions
# What
set up initial skeleton for request handlers. this included moving the echo (parse_data and the original response code) to a handler, and removing the tests pertaining to parse_data.

# Current behavior
currently echoes as expected for /echo, and responds with TODO your response for /static.

# Patchset 3
fixed some bugs with config parsing and added code to support parsing location/root

# TODO
- moving all config options into a struct (& related restructuring)
- file request handler - implementation should go in the file_request_handler.cc
- tests for echo request handler (move related session tests)
- eventual tests for file request handler

Change-Id: I06b5ac5bc87fc6df760b06a9802e0b1db662817a