Changed echo to specific paths and cleaned up logging/interface
#What
reflecting the Piazza posts, I've changed the echo to only echo configured paths and not every path not covered by static files. If a path is not covered, it will respond with 400 Bad Request. Also cleaned up the logging a bit and the interfaces (removing unused parameters) and changed tests to reflect the changes made.

#TODO
integration test for static file sometimes fails - probably trying to send response before it fully loads?

Change-Id: Ifb73638f8f7660b52c5840164636a509edaebfb2