made sure server is running, is closed down, and dockerfile edits for static file test
# What
added a loop to wait for server to come online locally, changed the kill line to target "webserver" in particular, and changed build part of Dockerfile to reflect actual paths on google cloud shell

# Why
previously, occasionally some tests would fail - this was due to attempting to send the command before server was up (running the server in the background caused them to run in parallel). Furthermore, while debugging, I realized that the webserver was never killed - I changed it to specifically target the webserver. Also, in many previous builds, the  local static file test would always break - I investigated and concluded that it is probably a Dockerfile issue - the original path was /project/ instead of /projects/raddish/ for instance, which made the local config (server_config) inaccurate.

# TODO
obviously still not 100% certain these are the issues (sometimes bash errors aren't obvious to me), please try running the Docker image locally first before approving

Change-Id: I603af529870d1a2fbed65ff2173baca48714c2a4