added support for longest match in config location paths
# What
If a config location is the substring of the request path and starts at index 0, take its mapped full path as the place to search for requested file.

# Why
initially, session parsed out the first section (before the first /) of the request path to match with the config location paths. This limited the config location paths to only one folder deep. By seeing if any of the config paths are a substring of the request path, it can be more flexible and take the longest path.

# Note
I also changed echo to respond in the same way - as the current config file has location "/" set to echo, that means that any other non-static-file paths will echo, for now. If this behavior is unintended, we can remove later.

Patchset edit: does longest match, whether in echo or static_file handler. Also made some other minor changes (e.g. single variable that holds the name of the handler instead of relying on two or more separate toggled booleans, minor logic bug fix, and pre-checking whether the http request is valid or not).


# TODO
better unit tests, probably after refactoring

Change-Id: I55f7d36182384eb3167fb00aa14b1349e7dc14d7