better refactoring
#What
Tried to separate out the functions better, and made a generic function to get specific parts of the config file to make future configs easier. Unfortunately I haven't figured out a good way to avoid writing fully custom config parsers for each section, as all the data they extract tends to be very specific to that piece of data.

As a note, "GetServerSettingsInner" is mostly a placeholder name - I want it to just be "GetServerSettings", but that's already the name of a function, and I can't change it until this code and @taffy's are merged as it's going to break all the refactored tests if I rename the function.

GetServerSettings will be renamed to something like "GetConfig" to reflect the more general nature of the function.

Change-Id: I22e29e076c54c16326210b4e244b5a4e22e784d1