config api change and removal of / as echo
# What
implemented a number of config api changes as dictated in the spec, and removed / as a path for echo

# Config API notes:
- already by default ignores # comments
- changed echo/static to handler names (EchoRequestHandler and FileRequestHandler)
- added support for quoting *around* strings (not inbetween, because parser doesn't tokenize it that way)
- added failing at startup if duplicate locations are found - made sure to double check whether trailing slashes existed (e.g. /text and /text/ should be considered duplicates)
- changed config output from static and echo file data structures to two mappings - (location, handler) and (location, root) (in preparation for factory) - if it's an echo it will not appear in the second mapping
- removed default root - not required and may be confusing

# Why
when the 404 handler is finished, it will be configured to /, Currently, / is configured to echo (which is what the integration test expects). If not changed, all the echo integration tests will receive 404 instead of 200. (copied from previous change, added to this one to avoid conflict)

# TODOs
- currently configs are still nested in a server block - not sure whether it's necessary to remove
- filesystem paths are relative, but not to webserver binary location (I think) - not entirely sure how to do this
- needs config unit tests for all of the new requirements

# Note
multiple changes were made to code/tests to accommodate for the API changes - they should be temporary and will be changed as the rest of the code is refactored

# Patches
- several changes to match config api more closely (e.g. port instead of listen)
- more logging
- fixed longest match issue (where some request like /echoNOTVALID would still match with /echo)

Change-Id: Id2b133f8c113bf0f75fcd75fe03a937a1dff46e3