echo handler now takes in req object
Change-Id: Ia082e23d9508ffb513b473785c21b7935264d893

added res objects

#What

This now does both res and req objects. Yay.

Some highlights:

Handlers now take in req objects and return res objects.
The base handler class now has its own constructor that auto-generates a default 404 response object to make future handlers easier, and takes in the request object and stores it so that methods can access the information in there. I might later switch this to a const or a reference, but it works for now.

File handler also takes in the root during construction.

Dispatcher also takes in the req and returns the res, instead of the previous strings.

We are now using http::async_write and http::async_read instead of asio. This has the nice side effect of removing the issue of not reading long headers in one pass, which means we don't need to do recursive calls anymore.

Sessions and handlers both store a res object now - sessions HAVE to do this, unfortunately, as otherwise we get a segfault when the write handler is called.

I've also added some better error logging so we can see error codes and such.

These changes should make writing additional handlers a lot easier, since we now have a more reasonable interface and the handlers auto-create responses and store the requests.


All surviving tests pass, including the integration ones, and random manual tests I've done work too.

#TODO
We need to re-write a ton of tests that I broke during this process.

Change-Id: I5d2da0ec166ebe12582d593fb5bfb65802eb365a