fixed handler and dispatcher tests, started on session tests
#What

With the req/res re-writes, changes to interface, etc, we had a lot of broken tests. This commit fixes all of the handler and dispatcher tests, and also fixes a couple of the session tests as an example of how to continue with those.

It also removes webserver from our coverage report (since that's just the main file to run the server in prod, and doesn't need testing), and gets us to ~94% coverage.

I also made a few bugfixes to the code, notably changing text/html to text/plain in 404 returns, making sure we had consistency in the 404 text.

Finally, I removed some old stuff we don't need anymore, like commented out functions and the old buffer vector we no longer use in the session.


#NOTE
I didn't have time to run the docker container - check that before you approve :)

Change-Id: I26e8129fca7cd61a6723400657f50f957cde5f53