Fixes factory to exit gracefully on incorrect handler config leading to no config found in map
#What
Some old code slipped into prod, where a random handler got selected if we couldn't find one in the map to deal with some weird static init issues.

This bug had been fixed, but lead to very weird behavior on a bad config where the handler is coded incorrectly and isn't added on startup, but config claims the handler does exist. It now does a fatal log and exits gracefully.


Change-Id: I0d3d8732e3e6f6af8f4a797fb4d26df3df3e93b6