Implement health request handler
#What
Added health request handler that returns HTTP 200 OK if called, configured to serve at /health

#Why
Per Assignment 8 spec, the server is simple enough that being able to return an OK response should be enough to consider the server healthy

#How
Followed README instructions on how to add a new request handler

#Tests
Wrote simple unit test checking that the handler returns HTTP 200 OK

#Note
We still have a lot of unneeded headers, so I tried to make my new files exclude those. In the future, we can use them as reference for which includes we don't need if we ever want to go through and prune unneeded includes again.

Change-Id: I44b259f48c4b93cbfe44ba6ae44b93efe14f9b22