Change making logs from BOOST_LOG_TRIVIAL to BOOST_LOG
Unless logs specify in log msg, it would be hard to tell what specifically made the logs. I wanted to make an attribute to tell so.
To do so, BOOST_LOG_TRIVIAL is too simple to do so, so system was switched to BOOST_LOG which can (tried to BOOST_LOG_SEV, but for some reason severity wouldnt show up)

- instead of BOOST_LOG_TRIVIAL(severity), use BOOST_LOG(lg)
- logging::sources::logger lg holds attributes to be set before making a log
- new attribute "Process" for describing what called log. should not change (server_main is exception). set when class is made / handler inits a response
- attribute "Severity" to determine what type log is. SHOULD be set before each log to make sure correct lvl (bc doesnt reset)

Change-Id: I95de1c57ff54ec74f2a7d0ac26ac8f9d964b23b8