Added basic game functionality
#What
Adds basic event loop and functions for creating/selling radishes, and a single upgrade.

#Details
Creates an event loop that ticks every second to add passive income, and fluctuate the market rate every 10 seconds for radish pricing.
Adds "buttons" in the index.html temporarily linked to functions for farming radishes, selling radishes, and buying farmers.
Implements a function that updates the text elements in the main display.


Change-Id: I0f0799f112dcda8cd207dfedde580dd7b2c97034