Create/Login/Logout
#What
Adds javascript and html for login/logout/create form. Also fixes the body duplication bug from sessions.

The login.html should be merged with our main index.html - the form can just be added, as it automatically centers itself, and hooked up with whatever login/logout buttons we end up using.

#How
The form now has two buttons - create and login. If you use create, it calls both create AND login, so that the player is automatically logged in on account creation. If you call login, it just pulls the data.

These scripts will probably have to be combined with the main js file, or the main js file will have to reference them, as they're the way to load player data.

Logout is currently broken due to a known issue - @tiffany should be handling that part.

#Testing
All tests pass still.


Change-Id: I034a9af7f7ad63302f4b9f325d3f329729d05fdf