Upgrades and money save/load properly now
#What
Adds save/load systems for upgrades/money, so the game now fully tracks progress.

#How
Everything is stored in upgrades.json in the database. In hindsight there's no reason to have a separate entry just for money, so I added it to "upgrades" even if it isn't technically one.

We now have a beforeunload listener that calls logout before closing the page. Note that this does get called even if you're not logged in, but it doesn't matter since it just causes an unseen 400 error (since the player won't be there to receive it anyway...)

Also, clears the database to make testing easier.

#TODO
Due to the lack of UI update functions, the UI doesn't actually update until you do something that would cause it to - for example, if you load, you will have 0 farmers but if you buy a farmer your old ones will "appear" (although you still get the effects of them beforehand).

Change-Id: I0483e039c99b4e050713f8dbbd8b065cc6afcdde